Made notifications more useful
I changed notifications to display the person who made the most recent
change along with their summary and description. If there's more than
one change it just shows the summary and for the description it states
how many more changes happened.

I also changed the build script for linux. At some point this needs to
look at what OS we're using and discover the chrome binary dynamically.
I added the pem into the repo and changed the script to use it so that
the plugin can just be updated rather than uninstalled and reinstalled.
Not like this will make it onto the Chrome store anyway so I'm not as
worried about security.